2.1.Installing Vue, Webpack & Eslint
To use Vue, we, of course, need to install it in our app. And, because we're using modern JavaScript practices, we're not going to include a script tag to a CDN or manually download Vue. We're going to install it with yarn.
But first, in addition to downloading Vue into our project, we also need to teach Webpack how to parse .vue files. Like React, Vue uses some special - not-actually-JavaScript - syntaxes, which Webpack needs to transform into real JavaScript.
To tell Webpack to parse vue files, open webpack.config.js. Near the bottom, though it doesn't matter where, add .enableVueLoader().
Yep! That's all you need. If you want to use Vue 3, you can pass an extra argument with a version key set to 3. Eventually, 3 will be the default version that Encore uses.
And, even though Encore is watching for changes, whenever you update webpack.config.js, you need to stop and restart Encore. I'll hit Control+C and then re-run:

yarn watch
When we do this... awesome! Encore is screaming at us! To use Vue, we need to install a few packages.