2.3.Installing Vue, Webpack & Eslint{What is Vue?; Creating a Targwt Element for Vue; Our non-single application}
What is Vue?
In the simplest sense, Vue is a templating engine written in JavaScript. That over-simplifies it... but it's more or less true. In Symfony, if you go back to ProductController, we're accustomed to using Twig. It's easy: we tell it what template to render and we can pass variables into that template.
The Twig file itself is just HTML where we have access to a few Twig syntaxes, like {{ variableName }} to print something.
Vue works in much the same way: instead of Twig rendering a template with some variables, Vue will render a template with some variables. And the end-result will be the same: HTML. Of course, the one extra super power of Vue is that you can change the variables in JavaScript, and the template will automatically re-render.

Creating a Target Element for Vue
So instead of rendering this markup in Twig, delete all of it and just add <div id="app">. That id could be anything: we're creating an empty element that Vue will render into.

Our Non-Single Page Application
Now, what we're building will not be a single page application, and that's on purpose. Using Vue or React inside of a traditional web app is actually trickier than building a single page application. On our site, the homepage will soon contain a Vue app... but the layout - as you can see - is still rendered in Twig. We also have a login page which is rendered completely with Twig and a registration page that's the same. We'll purposely use Vue for part of our site, but not for everything... at least not in this tutorial.