2.4.Installing Vue, Webpack & Eslint{Creating a Second Webpack Entry}
Creating a Second Webpack Entry
Go back and open webpack.config.js again. This has one entry called app. The purpose of app is to hold any JavaScript or CSS that's used across our entire site, like to power the layout. We actually don't have any JavaScript, but the app.scss contains the CSS for the body, header and other things. The app script and link tags are included on every page.
But, our Vue app isn't going to be used on every page. So instead of adding our code to app.js, let's create a second entry and include it only on the pages that need our Vue app.
Copy the first addEntry() line, paste, and rename it to products - because the Vue app will eventually render an entire product section: listing products, viewing one product and even a cart and checkout in the next tutorial.
Now, in assets/js, create that file: products.js. Let's start with something exciting: a console.log():

  Boring JavaScript file: make me cooler!