2.6.Installing Vue, Webpack & Eslint{Our First Vue Instance}
Our First Vue Instance
Let's... make it cooler with Vue! Back in products.js, start by importing Vue: import vue from 'vue'. This is one of the few parts that will look different in Vue 3 - but the ideas are the same.
If you imagine that Vue is a templating engine - like Twig - then all we should need to do is pass Vue some template code to render. And... that's exactly what we're going to do. Add const app = new Vue() and pass this some options. The first is el set to #app. That tells Vue to render inside of the id=app element. Then, pass one more option: template. This is the HTML template - just like a Twig template - except that, for now, we're going to literally add the HTML right here, instead of in a separate file:

  <h1>Hello Vue! Is this cooler?</h1>

That's... all we need! Moment of truth: find your browser and refresh. There it is! We just built our first Vue app in about 5 lines of code.
Next, let's make it more interesting by passing variables to the template and witnessing Vue's awesomeness first hand.