3.1.Vue Instance & Dynamic Data{The data Option}
We've just seen the most basic thing you can do with Vue. And if you think of Vue as a templating engine like Twig, it makes a lot of sense: we instantiated a new Vue instance, told it where on the page to render and passed it a template. And that totally worked. Booya!

The data Option
When you instantiate Vue, you control it by passing a number of different options... and a lot of this tutorial will be about learning what options are possible. One of the most important ones is data. Unlike el and template, data is a function. It returns an array - or, really, this is an "object" in JavaScript - of variables that you want to pass into the template.
Notice that ESLint is angry - it's because this line is empty. Sometimes you need to ignore it until you finish: it's a bit overeager. Let's create one new "data" - one new "variable" - to pass into the template: firstName set to Ryan. That's me!
And now that we're passing a firstName variable into the template, we can say "Hello" and {{ firstName }}. Yes, by complete coincidence, Vue uses the same syntax as Twig to render things.
Before we try this, ESLint is still mad at me. Sheesh! It says:

  Expected method shorthand