3.2.Vue Instance & Dynamic Data{The data Option}
As I mentioned, some of the options you pass to Vue are set to values - like el and template, while others are functions. When you have a method in an object like this, you can use a shorthand: data() {... which is just a lot more attractive.
Oh, and at the bottom, temporarily add window.app = app. That will set our Vue app as a global variable, which will let us play with it in our console. Ready?
Refresh! It rendered! But it gets better! In your browser's console, type app.firstName. You can already see that this equals Ryan! Any data key becomes accessible as a property on our instance. Set this to Beckett - my son's name... who is hopefully napping right now. Boom! The template immediately updates for the new data. And we can change this over and over again.
So Vue is a lot like Twig - it renders templates and can pass variables into the templates - but with this crazy-cool extra power that when we change a piece of data, it automatically re-renders... which, of course, is exactly what we want.