3.3.Vue Instance & Dynamic Data{How the Vue Instance Rendering Really Works}
How the Vue Instance Rendering Really Works
And... at a high level... that's Vue! Yes, we're going to talk about so much more, but you already understand its main purpose. Remove the global variable and the app = code - we don't need that. ESLint will temporarily get mad because it thinks it's weird that we're instantiating an object and not setting to a variable, but that's fine... and it'll go away in a little while.
Behind the scenes, when Vue renders, it actually calls a render() method on the object, which you don't normally need to worry or care about. But to help this all make more sense, I want you to see what this method looks like one time. Stick with me, we're going to do some temporary experimentation.
I'm going to add a new render() function. As soon as I add this, when Vue renders it will call our function instead of rendering it internally for us. But inside, I'm going to put the exact code that Vue normally runs: return Vue.compile(this.$options.template) - that's a special way to reference the template option here - .render.call(this, h).
I know, that's totally crazy, and you will never need to type this in a real project. What this shows us is that Vue has a compile() function where you can "compile" a template string and then call render() on it. The .render.call thing is a fancy way of basically calling .render() and passing it the h variable, which is another object that's good at dealing with DOM elements... and that you don't need to worry about.
If we refresh now, it works exactly like before because our render() method does exactly what Vue normally does.