3.4.Vue Instance & Dynamic Data{How the Vue Instance Rendering Really Works}
If you look at this, you might start to wonder: why do we need a template option at all? We're just reading it in render()... so it could live anywhere. Let's try that! Remove the option, and, at the top, add const template = the template string. In render, reference that local variable.
That should work, right? Let's find out. It totally does!
Ok, so let me tell you the big important point I'm trying to make. Vue... is simple: it's a system where you can take a template string, render it, and pass this data into the template... just like Twig.
Of course, as we start adding more to our app, this template variable is going to get huge and ugly. To help with that, Vue has a very special organizational concept called single file components. Let's create one next.