4.3.Single File Component{Adding a Component Name; $mount() the Component}
Adding a Component Name
From here on out, we're going to do pretty much all our work inside of these .vue files - called single file components. One option that we're going to add to every component is name: set it to Products. We could use any name here: the purpose of this option is to help debugging: if we have an error, Vue will tell us that it came from the Products component. So, always include it, but it doesn't change how our app works.

$mount() the Component
Before we keep working, there are two small changes I want to make to products.js. First, the el option: it tells Vue that it should render into the id="app" element on the page. This works, but you usually see this done in a different way. Remove el and, after the Vue object is created, call .$mount() and pass it #app. I also like this better: we first create this Vue object - which is a template and set of data that's ready to go - and then choose where to mount it on the page.