4.4.Single File Component{Shorthand render() Method}
Shorthand render() Method
Second, because the render() method only contains a return line, we can shorten it: render set to h => h(App).
That's effectively the same: it uses the arrow function to say that render is a function that accepts an h argument and will return h(App). I'm mostly making this change because this is how you'll see Vue apps instantiated on the web.
Next, let's get to work inside our single file component: we'll add the HTML markup needed for our product list page and then learn how we can add styles.