6.2.data() and Vue Dev Tools{ The data() Function versus data: () => { Function }
The data() Function versus data: () => { Function
Before we create our second component, I need to point out a small detail. We configure Vue by passing it options. Some of these options are set directly to values, while others - like data() - are functions.
And because the data() function is always just a return statement, you'll often see it written with the shortcut, arrow syntax: data: (), arrow, ({, remove the return and fix the ending.
Just like with the render shortcut we used in products.js, this is effectively the same: it says that the data property is set to a function that returns this object. The return is implied.