6.3.data() and Vue Dev Tools{ The data() Function versus data: () => { Function }
You can use this if you want... but I'm going to go back to the original way. It's a bit longer, but I'll use this syntax consistently for all Vue options that are set to functions. The shorter syntax also has a limitation where you can't use the this variable: something that we will need later.
Next: let's extract part of our product listing markup into a second component and include it from Products. This will start to show off one of Vue's key concepts: having multiple components that communicate to each other.