9.1.v-bind: Dynamic Attributes
We added a title prop to legend to make it dynamic: whenever we use this component, we can pass it different text. It's like an argument to the legend component.
But in this situation, I want to go one step further: I want to pretend that the legend text on this page needs to change while our app is running - like the shipping time magically starts to decrease if the user is on the site for awhile... and we really want them to buy.
Whenever we need a value to change while the app is running, that value needs to be stored as data. In products.vue, we already have a legend data from earlier, but we're temporarily not using it. So what we really want to do is this: pass the legend data as the title prop, instead of the hardcoded text.
Easy enough! We know that anything in data and props is available in our template. So, for the title attribute, or technically prop, say title="{{ legend }}"
As soon as we do that, Webpack is mad! Go check out the terminal. Yikes, it doesn't like this syntax at all - it can't even build our assets:

  interpolation inside attributes has been removed. Use v-bind or the colon shorthand instead. For example, instead of id="{{ val }}", use :id="val".

That's... a pretty awesome error message.