9.2.v-bind: Dynamic Attributes{Using v-bind for Dynamic Attributes}
Using v-bind for Dynamic Attributes
Basically, the {{ }} syntax that we've grown to know and love... can't be used inside an attribute. If you need a dynamic value in an attribute, you need to prefix the attribute with v-bind:. And then, inside the attribute, just use the variable name.
Now it builds successfully. Before we talk more about this, let's try it! Refresh and... it works! Over on the Vue dev tools, the Products component has a legend data... and it looks like the Legend component is receiving that as its title prop.
The cool thing is that if we modify the legend data - "Will ship slowly! " - the text updates! The modified data is passed to Legend as the title prop, and Vue re-renders it. So while we will never change a prop directly, if we change a data... and that data is passed to a prop, the prop will update to reflect that.