9.3.v-bind: Dynamic Attributes{v-bind is Full JavaScript}
v-bind is Full JavaScript
Back at our editor, let's talk more about this v-bind thing. There will actually be several of these v- things in Vue: they're used whenever Vue needs to do something special, including if statements and for loops. v-bind is probably the most important one. Very simply: if you want an attribute to be set to a dynamic value, you must prefix the attribute with v-bind.
As soon as you do that, the attribute is no longer just text, like pb-2 up here. We're now writing JavaScript inside the attribute.

I'll prove it: add + open quote and say

  this is really JavaScript.

And... when we try this... yea! That text shows up! It's a mixture of our data and that string.
So... that's really it! v-bind is meant to "bind" an attribute to some dynamic value, often a data or prop. But honestly, I don't even think of it like that. I just think: if I want to use JavaScript inside of an attribute, I need to use v-bind.