9.4.v-bind: Dynamic Attributes{The v-bind Colon Shorthand; Specifying the type of a Prop}
The v-bind Colon Shorthand
We're going to use this all the time: we're constantly going to be setting attributes to dynamic values. Vue understand this. And so, they've provided us with a nice shortcut. Instead of v-bind:title, just say :title.
That means the exact same thing: it's still really v-bind behind the scenes.
I like this way more. In my mind, an attribute without a colon is literally set to that string. Prefixing the attribute with : transforms it into JavaScript. Simple.
Anyways, when we try it now it... of course, works brilliantly.

Specifying the type of a Prop
While we're talking about the legend component, I want to make one small tweak to the props option. Remember: to allow a title prop to be passed to us, we needed to first define the prop here.

If you hover over this, ESLint is mad:

  prop title should define at least its type.

In addition to just saying:

  please allow a prop called title to be passed to me

We can also tell Vue what type we expect this prop to be and whether or not it's required.