10.1.Modular CSS{CSS Name Conflicts; Hello Modular CSS}
I love that I can put my styles right inside the component. In products.vue, we render an element with a sidebar class... and then immediately - without going anywhere else - we're able to add the CSS for that. We can still have external CSS files with shared CSS, but for any styling that's specific to a component, it can live right there.

CSS Name Conflicts
But... we need to be careful. The class sidebar is a pretty generic name. If we accidentally add a sidebar class to any other component - or even to an element in our Twig layout - this CSS will affect it!
Find your browser, refresh, and open the HTML source. On top, here's our stylesheet: /build/products.css. Open that up. Yep! That's what I expected: a .sidebar class with the CSS from the component. It's easy to see that these styles would affect any element on the page with a sidebar class... whether we want it to or not.

Hello Modular CSS
To solve this problem, Vue, well really, the CSS world - because this is a generic CSS problem - has created two solutions: scoped CSS and modular CSS. They're... two slightly different ways to solve the same problem and you can use either inside of Vue. We're going to use modular CSS.
What does that mean? It means that whenever we have a style tag in a Vue component, we're going to include a special attribute called module.
That's it. Back at your browser, leave the CSS file open, but close the HTML source and refresh the homepage. Ah! We lost our sidebar styling! It's a modular CSS "feature", I promise!
To see what's going on, go back to the tab with the CSS file and refresh. Woh. The class names changed: they're now .sidebar_ and then a random string. Back on the main tab, if we inspect element... the div still has the normal sidebar class.
Here's what's going on. When you add module to the style tag, Vue generates a random string that's specific to this component and adds that to the end of all class names. The great thing is that we can use generic class names like sidebar without worrying about affecting other parts of our page. Because... in the final CSS, the class name is really sidebar_ then that random string.
Of course, now that this is happening, we can't just say class="sidebar" in the template anymore. We need to somehow use the dynamic name - the one that includes the random string.