10.2.Modular CSS{Rendering the Dynamic CSS Class}
Rendering the Dynamic CSS Class
As soon as you add module to a style tag, Vue makes a new variable available in your template called , which is a map from the original class name - like sidebar - to the new dynamic name - like sidebar_abc123.
I'm going to delete the p-3 mb-5 classes temporarily... just to simplify.
Ok: we no longer want to set the class attribute to a simple string: it needs to be dynamic. And whenever an attribute needs to contain a dynamic value, we prefix it with :, which is really v-bind dressed up in its superhero costume.
Now, we're writing JavaScript. Use that new $style variable to say $style.sidebar.
Unfortunately, at this time, the Vue plugin in PhpStorm doesn't understand the $style variable, so it won't be much help here. But because we have a class called sidebar inside the style tag, we can say $style.sidebar.
Let's try it! Move over, refresh and... we're back! Our class renders with the dynamic name.
Of course, it looks a little weird because our element is missing the two classes we removed. How can we add them back? We could do some ugly JavaScript, a plus, quote, space... but... come on! Vue almost always has a nicer way.